+++ jarvis.py
@@ -1,194 +1,35 @@
-import os
 import pathlib
-import string
-import struct
-import sys
 import time
-from datetime import datetime
+import warnings
 from multiprocessing import Manager, Process
-from typing import Dict, NoReturn, Union
 
-import pvporcupine
-from pyaudio import PyAudio, Stream, paInt16
-
-from modules import listener, speaker
-from modules.api_handler import make_request
-from modules.config import config, time_converter
+from executables.helper import time_converter, word_engine
+from modules.exceptions import APIError
 from modules.logger import logger
-from modules.models import env, fileio, flag, settings
-from modules.playsound import playsound
+
+try:
+    from multiprocessing.managers import DictProxy  # noqa
+except ImportError:
+    from typing import Dict as DictProxy
 
 
-def processor() -> Union[str, None]:
-    """Processes the request after wake word is detected.
-
-    Returns:
-        bool:
-        Returns a ``True`` flag if a manual stop is requested.
-    """
-    if phrase := listener.listen():
-        logger.info(f"Request: {phrase}")
-        sys.stdout.write(f"\rRequest: {phrase}")
-        if "restart" in phrase.lower():
-            logger.info("User requested to restart.")
-            playsound(sound=fileio.restart)
-            return flag.restart
-        if "stop running" in phrase.lower():
-            logger.info("User requested to stop.")
-            playsound(sound=fileio.shutdown)
-            return flag.stop
-        if not any(word in phrase.lower() for word in config.keywords + config.conversation):
-            logger.warning(f"'{phrase}' is not a part of recognized keywords or conversation.")
-            return
-        if not any(word in phrase.lower() for word in config.api_compatible['compatible']):
-            logger.warning(f"'{phrase}' is not a part of API compatible request.")
-            playsound(sound=fileio.unprocessable)
-            return
-        if any(word in phrase.lower() for word in config.delay_with_ack + config.delay_without_ack):
-            logger.info(f"Increasing timeout for: {phrase}")
-            timeout = 30
-            if any(word in phrase.lower() for word in config.delay_with_ack):
-                playsound(sound=fileio.processing, block=False)
-        else:
-            timeout = env.request_timeout
-        if response := make_request(path='offline-communicator', timeout=timeout,
-                                    data={'command': phrase, 'native_audio': env.native_audio,
-                                          'speech_timeout': env.speech_timeout}):
-            if response is True:
-                logger.info("Response received as audio.")
-                sys.stdout.write("\rResponse received as audio.")
-                playsound(sound=fileio.speech_wav_file)
-                os.remove(fileio.speech_wav_file)
-                return
-            response = response.get('detail', '')
-            logger.info(f"Response: {response}")
-            sys.stdout.write(f"\rResponse: {response}")
-            speaker.speak(text=response)
-        else:
-            playsound(sound=fileio.failed)
-
-
-class Activator:
-    """Awaits for the keyword ``Jarvis`` and triggers ``initiator`` when heard.
-
-    >>> Activator
-
-    See Also:
-        - Creates an input audio stream from a microphone, monitors it, and detects the specified wake word.
-        - Once detected, Jarvis triggers the ``listener.listen()`` function with an ``acknowledgement`` sound played.
-        - After processing the phrase, the converted text is sent as response to the API.
-    """
-
-    def __init__(self):
-        """Initiates Porcupine object for hot word detection.
-
-        See Also:
-            - Instantiates an instance of Porcupine object and monitors audio stream for occurrences of keywords.
-            - A higher sensitivity results in fewer misses at the cost of increasing the false alarm rate.
-            - sensitivity: Tolerance/Sensitivity level. Takes argument or env var ``sensitivity`` or defaults to ``0.5``
-
-        References:
-            - `Audio Overflow <https://people.csail.mit.edu/hubert/pyaudio/docs/#pyaudio.Stream.read>`__ handling.
-        """
-        keyword_paths = [pvporcupine.KEYWORD_PATHS[x] for x in env.wake_words]
-        self.py_audio = PyAudio()
-        arguments = {
-            "library_path": pvporcupine.LIBRARY_PATH,
-            "sensitivities": env.sensitivity
-        }
-        if settings.legacy:
-            arguments["keywords"] = env.wake_words
-            arguments["model_file_path"] = pvporcupine.MODEL_PATH
-            arguments["keyword_file_paths"] = keyword_paths
-        else:
-            arguments["model_path"] = pvporcupine.MODEL_PATH
-            arguments["keyword_paths"] = keyword_paths
-
-        self.detector = pvporcupine.create(**arguments)
-        self.audio_stream = self.open_stream()
-        label = ', '.join([f'{string.capwords(wake)!r}: {sens}' for wake, sens in
-                           zip(env.wake_words, env.sensitivity)])
-        self.label = f"Awaiting: [{label}]"
-
-    def __del__(self) -> NoReturn:
-        """Invoked when the run loop is exited or manual interrupt.
-
-        See Also:
-            - Releases resources held by porcupine.
-            - Closes audio stream.
-            - Releases port audio resources.
-        """
-        self.detector.delete()
-        if self.audio_stream and self.audio_stream.is_active():
-            self.py_audio.close(stream=self.audio_stream)
-            self.audio_stream.close()
-        self.py_audio.terminate()
-
-    def open_stream(self) -> Stream:
-        """Initializes an audio stream.
-
-        Returns:
-            Stream:
-            PyAudio stream.
-        """
-        return self.py_audio.open(
-            rate=self.detector.sample_rate,
-            channels=1,
-            format=paInt16,
-            input=True,
-            frames_per_buffer=self.detector.frame_length,
-            input_device_index=env.microphone_index
-        )
-
-    def executor(self, status):
-        """Closes the audio stream and calls the processor."""
-        logger.debug(f"Detected {settings.bot} at {datetime.now()}")
-        status["LOCKED"] = True
-        logger.debug("Restart locked")
-        playsound(sound=fileio.acknowledgement, block=False)
-        self.py_audio.close(stream=self.audio_stream)
-        processed = processor()
-        if processed == flag.stop:
-            self.audio_stream = None
-            raise KeyboardInterrupt
-        if processed == flag.restart:
-            status["LOCKED"] = None
-        else:
-            status["LOCKED"] = False
-        logger.debug("Restart released")
-        self.audio_stream = self.open_stream()
-
-    def start(self, status: Dict) -> NoReturn:
-        """Runs ``audio_stream`` in a forever loop and calls ``initiator`` when the phrase ``Jarvis`` is heard."""
-        logger.info(f"Starting wake word detector with sensitivity: {env.sensitivity}")
-        while True:
-            sys.stdout.write(f"\r{self.label}")
-            pcm = struct.unpack_from("h" * self.detector.frame_length,
-                                     self.audio_stream.read(num_frames=self.detector.frame_length,
-                                                            exception_on_overflow=False))
-            result = self.detector.process(pcm=pcm)
-            if settings.legacy:
-                if len(env.wake_words) == 1 and result:
-                    settings.bot = env.wake_words[0]
-                    self.executor(status=status)
-                elif len(env.wake_words) > 1 and result >= 0:
-                    settings.bot = env.wake_words[result]
-                    self.executor(status=status)
-            else:
-                if result >= 0:
-                    settings.bot = env.wake_words[result]
-                    self.executor(status=status)
-
-
-def starter(status: Dict) -> None:
+def initiator(status_manager: DictProxy) -> None:
     """Starts main process to activate Jarvis and process requests via API calls."""
     try:
-        Activator().start(status=status)
-    except KeyboardInterrupt:
-        return
+        # Import within a function to catch startup errors
+        from executables.starter import Activator
+    except APIError as error:
+        logger.error(error)
+        status_manager["LOCKED"] = "RESTART"
+    else:
+        try:
+            status_manager["LOCKED"] = False
+            Activator().start(status_manager=status_manager)
+        except KeyboardInterrupt:
+            return
 
 
-def terminate(process: Process):
+def terminator(process: Process):
     """Terminates the process.
 
     Args:
@@ -216,32 +57,44 @@ def begin():
         - Reloads env variables upon restart.
         - Avoids memory overload.
     """
-    sys.stdout.write(f"\rRestart set to {time_converter(second=env.restart_timer)}")
+    # Import within a function to be called repeatedly
+    from modules.models import env
     logger.info(f"Restart set to {time_converter(second=env.restart_timer)}")
-    status_dict = Manager().dict()
-    status_dict["LOCKED"] = False
-    process = Process(target=starter, args=(status_dict,))
+    status_manager: DictProxy = Manager().dict()
+    status_manager["LOCKED"] = False  # Instantiate DictProxy
+    process = Process(target=initiator, args=(status_manager,))
     process.name = pathlib.Path(__file__).stem
     process.start()
     logger.info(f"Initiating as {process.name}[{process.pid}]")
     start_time = time.time()
     while True:
         if start_time + env.restart_timer <= time.time():
-            if status_dict["LOCKED"]:
+            if status_manager["LOCKED"] is True:
                 continue  # Wait for the existing task to complete
             logger.info(f"Time to restart {process.name}[{process.pid}]")
-            terminate(process=process)
+            terminator(process=process)
             break
         if not process.is_alive():
+            if begin.count > env.restart_attempts:
+                warnings.warn(
+                    "Retry limit exceeded after consecutive start up errors."
+                )
+                logger.critical("Retry limit exceeded.")
+                return
+            if status_manager["LOCKED"] == "RESTART":  # Called only when restart fails during initial connect
+                begin.count += 1
+                logger.warning(f"{word_engine.ordinal(begin.count)} restart because of a problem.")
+                break
             logger.info(f"Process {process.name}[{process.pid}] died. Ending loop.")
             return
-        if status_dict["LOCKED"] is None:
+        if status_manager["LOCKED"] is None:
             logger.info("Lock status was set to None")
-            terminate(process=process)
+            terminator(process=process)
             break
         time.sleep(0.5)
     begin()
 
 
 if __name__ == '__main__':
+    begin.count = 0
     begin()